refactor(expenses): use a catch-all api route
